refactor: 로그인 UI 작업 모듈화 및 수정
