--- NangPaGo-fe/src/common/utils/validation.js
@@ -1,5 +1,0 @@
-import { VALID_PROVIDERS } from '../constants/providers';
-
-export const isValidProvider = (provider) => {
-  return VALID_PROVIDERS.includes(provider);
-};
