--- NangPaGo-fe/src/components/SocialLoginButton/utils.js
@@ -1,6 +1,0 @@
-export const getStyleForProvider = (styles, provider) => {
-  if (!styles[provider]) {
-    throw new Error(`알 수 없는 제공자: ${provider}`);
-  }
-  return styles[provider];
-};
