+++ NangPaGo-fe/src/components/util/auth.js
@@ -1,15 +1,15 @@
 const SOCIAL_BUTTON_STYLES = {
-  naver: {
-    background: 'bg-[#03C75A] hover:bg-[#02B053]',
-    text: 'text-white',
-    logo: '/socialLogin/naver.svg',
-    label: '네이버 계정으로 로그인',
-  },
   google: {
     background: 'bg-white border border-[#DADCE0] hover:bg-[#F6F6F6]',
     text: 'text-[#3C4043]',
     logo: '/socialLogin/google.svg',
     label: '구글 계정으로 로그인',
+  },
+  naver: {
+    background: 'bg-[#03C75A] hover:bg-[#02B053]',
+    text: 'text-white',
+    logo: '/socialLogin/naver.svg',
+    label: '네이버 계정으로 로그인',
   },
   kakao: {
     background: 'bg-[#FEE500] hover:bg-[#FFEE32]',
