+++ NangPaGo-fe/src/main.jsx
@@ -1,10 +1,10 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App.jsx';
 import './globals.css';
+import Router from './pages/index.jsx';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    <Router />
   </StrictMode>,
 );
